Move Change service and graphql into ChangeLogs
