+++ frontend/src/service/ChangeLog.service.ts
@@ -73,4 +73,19 @@ export class ChangeLogService {
 
     return changes;
   }
+
+  createChange(title: string, summaryMarkdown: string): Promise<Change> {
+    return new Promise<Change>((resolve, reject) => {
+      this.changeLogGraphQLApi
+        .createChange(title, summaryMarkdown)
+        .then(resolve)
+        .catch(errCode => {
+          if (errCode === Err.Unauthenticated) {
+            reject({ authenticationErr: 'User is not authenticated' });
+            return;
+          }
+          reject({ changeErr: this.errorService.getErr(errCode) });
+        });
+    });
+  }
 }
